requirements.txt
pytesseract
pdf2image
pillow
pdfplumber
fastapi
uvicorn
python-multipart
pymongo
motor
pypdf
python-docx
python-dotenv
openai
sentence-transformers
